feat(*): COM-2735 rename page components folders
