+++ components/design-lab/Hero/index.js
